firmware: at_cmd: use "TwoCC" mechanism for the AT command codes
The internal AT code is just the ASCII representation of the command
stuffed into a 16-bit value.

This makes conversion between the code and a string trivial and a the same
time it allows us to work with the code only defined in one place at the
`at_cmds[]` table without too much macro magic.

Signed-off-by: Martin Pietryka <martin@pietryka.at>