firmware: drivers: lis2mdl: switch SPI mode correctly to 4 wire
The `lis2mdl_spi_mode_set()` function performs a read-modify-write, however
if we are not in 4 wire mode yet, then reading over SPI might give us 0xFF
so the function will write 0xFF to the register, enabling things like the
self-test mode which is bad.

Instead just write the desired value directly to the register.

Signed-off-by: Martin Pietryka <martin@pietryka.at>